Add context manager protocol
This allows to create a new PyXmv object with the syntax

```
with PyXmv(...) as pyxmv:
    ...
```

This guarantees that the underlying NuXmv process is killed right after
the interpreter leaves the with-block. This, in turn immediately frees
up all memory used by nuxmv. One can still use the "normal" syntax
```
pyxmv = PyXmv(...)
```
But then the nuxmv process is only killed when the GC collects the
pyxmv object, which might take longer.